JC: Finish up installation guide

+++ scripts/deploy.sh
@@ -3,6 +3,12 @@
 dots=$HOME/Documents/Git/dotfiles
 config=$HOME/.config
 
+if [ "$1" = "--env" ]; then
+    ln -fs $dots/zsh/zshenv $config/zsh/.zshenv
+    ln -fs $dots/zsh/zshenv $HOME/.xprofile
+    exit
+fi
+
 rm -rf $config/bspwm   2>/dev/null
 rm -rf $config/latexmk 2>/dev/null
 rm -rf $config/npm     2>/dev/null
@@ -38,11 +44,10 @@ ln -fs $dots/vim/coc-settings.json   $config/nvim/
 ln -fs $dots/vim/vimrc               $config/nvim/init.vim
 ln -fs $dots/zsh/zlogout             $config/zsh/.zlogout
 ln -fs $dots/zsh/zshrc               $config/zsh/.zshrc
-ln -fs $dots/zsh/zshenv              $config/zsh/.zshenv
-ln -fs $dots/zsh/zshenv              $HOME/.xprofile
 ln -fs $dots/scripts/p.js            $HOME/.local/bin/p
 
 cp $dots/font/JetBrains_Mono_Regular_Nerd_Font_Complete_Mono.ttf ~/.local/share/fonts/
 sudo fc-cache -fv
 
+echo
 echo "Finished deployment"
